fix endpoints which involve uploading a file.
also renames "requestWithValueBody" to "requestWithJsonBody", as well
as audit log message support for setupMultipartRequest.
noteworthy that endpoints which allow uploading files now take
`rest.Upload` instead of `*std.Io.Reader`, which requires setting the
filename and content-type of the file, as well as checking for file sizes to allow
the content-length header to be set.